--- agimus_demo_01_lfc_alone/config/controller_manager.yaml
@@ -1,15 +1,0 @@
-controller_manager:
-  ros__parameters:
-    update_rate: 1000  # Hz
-
-    linear_feedback_controller:
-      type: linear_feedback_controller/LinearFeedbackController
-
-    joint_state_estimator:
-      type: linear_feedback_controller/JointStateEstimator
-
-    joint_state_broadcaster:
-      type: joint_state_broadcaster/JointStateBroadcaster
-
-    franka_robot_state_broadcaster:
-      type: franka_robot_state_broadcaster/FrankaRobotStateBroadcaster
