+++ agimus_demo_03_mpc_dummy_traj/launch/bringup.launch.py
@@ -1,12 +1,15 @@
 from launch import LaunchContext, LaunchDescription
-from launch.actions import OpaqueFunction
+from launch.actions import OpaqueFunction, RegisterEventHandler
 from launch.launch_description_entity import LaunchDescriptionEntity
 from launch_ros.substitutions import FindPackageShare
+from launch.substitutions import PathJoinSubstitution
 from launch_ros.actions import Node
+from launch.event_handlers import OnProcessExit
 
 from controller_manager.launch_utils import (
-    generate_controllers_spawner_launch_description,
+    generate_load_controller_launch_description,
 )
+
 
 from agimus_demos_common.launch_utils import (
     generate_default_franka_args,
@@ -19,17 +22,24 @@ def launch_setup(
 ) -> list[LaunchDescriptionEntity]:
     franka_robot_launch = generate_include_franka_launch()
 
-    lfc_params = PathJoinSubstitution(
+    linear_feedback_controller_params = PathJoinSubstitution(
         [
             FindPackageShare("agimus_demo_03_mpc_dummy_traj"),
             "config",
-            "lfc_params.yaml",
+            "linear_feedback_controller_params.yaml",
         ]
     )
 
-    load_joint_state_estimator = generate_controllers_spawner_launch_description(
-        controller_names=["joint_state_estimator", "linear_feedback_controller"],
-        controller_params_file=lfc_params,
+    load_linear_feedback_controller = generate_load_controller_launch_description(
+        controller_name="linear_feedback_controller",
+        controller_params_file=linear_feedback_controller_params,
+        extra_spawner_args=["--controller-manager-timeout", "1000"],
+    )
+
+    load_joint_state_estimator = generate_load_controller_launch_description(
+        controller_name="joint_state_estimator",
+        controller_params_file=linear_feedback_controller_params,
+        extra_spawner_args=["--controller-manager-timeout", "1000"],
     )
 
     agimus_controller_yaml = PathJoinSubstitution(
@@ -48,18 +58,22 @@ def launch_setup(
         parameters=[agimus_controller_yaml],
     )
 
-    simple_trajectory_publisher_node = (
-        Node(
-            package="agimus_controller_ros",
-            executable="simple_trajectory_publisher",
-            name="simple_trajectory_publisher",
-            output="screen",
-        ),
+    simple_trajectory_publisher_node = Node(
+        package="agimus_controller_ros",
+        executable="simple_trajectory_publisher",
+        name="simple_trajectory_publisher",
+        output="screen",
     )
 
     return [
         franka_robot_launch,
-        load_joint_state_estimator,
+        load_linear_feedback_controller,
+        RegisterEventHandler(
+            event_handler=OnProcessExit(
+                target_action=load_linear_feedback_controller.entities[-1],
+                on_exit=[load_joint_state_estimator],
+            )
+        ),
         agimus_controller_node,
         simple_trajectory_publisher_node,
     ]
